+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 0.36 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Add Transifex.net service integration to manage the translation process.
+  [macagua]
+- Add Spanish translation
+  [macagua]
 
 
 0.35 (2020-11-18)